Index

This page lists the core methods and types of the Julia component of the DCCR project.

Index

This section enumerates the names exported by the package, each of which links to its corresponding Documentation.

Modules

DCCR.DCCR

Methods

Types

Constants

DCCR.DCCR_VERSION

Docs

This section lists the documentation for every exported name of the DCCR package.

Modules

DCCR.DCCR — Module

A module encapsulating the experiment driver code for the DCCR project.

Imports

The following names are imported by the package as dependencies:

AdaptiveResonance

ArgParse

Base

ColorSchemes

Core

DataFrames

DataStructures

Dates

DelimitedFiles

DocStringExtensions

DrWatson

JLD2

JSON

Logging

MLBase

MLDataUtils

NumericalTypeAliases

Pkg

Plots

Printf

ProgressMeter

PyCall

Random

Reexport

StatsPlots

Exports

The following names are exported and available when using the package:

DCCR_VERSION

Functions

Types

Constants

DCCR.DCCR_VERSION — Constant

A constant that contains the version of the installed DCCR package.

This value is computed at compile time, so it may be used to programmatically verify the version of OAR that is installed in case a compat entry in your Project.toml is missing or otherwise incorrect.