Developer Index

This page lists the types and functions that are internal to the DCCR package. Because they are not part of the public API, these names might change relatively frequently between versions and so should not be relied upon.

All internal names are listed in the Index, and each of these entries link to the docstrings in the Docs section.

Index

This section contains a list of internal names that link to their corresponding Documentation.

Methods

DCCR.evaluate_agent!

DCCR.fields_to_dict!

DCCR.get_index_from_name

DCCR.initialize_exp_queue!

DCCR.is_complete

DCCR.json_load

DCCR.json_save

DCCR.log_data

DCCR.run_scenario

DCCR.sanitize_block_type

DCCR.sanitize_in_list

DCCR.sanitize_log_state

Types

DCCR.Agent

DCCR.DDVFAAgent

DCCR.DDVFAAgent

DCCR.DDVFAAgent

DCCR.Data

DCCR.DataSplit

DCCR.DataSplitCombined

DCCR.DataSplitCombined

DCCR.DataSplitIndexed

DCCR.Experience

DCCR.Experience

DCCR.ExperienceQueue

DCCR.ExperienceQueueContainer

DCCR.ExperienceQueueContainer

DCCR.ExperienceQueueContainer

DCCR.Features

DCCR.LabeledDataset

DCCR.LabeledDataset

DCCR.Labels

DCCR.MatrixData

DCCR.SequenceNums

DCCR.StatsDict

DCCR.Targets

DCCR.VectorLabeledDataset

DCCR.VectoredData

Constants

DCCR.BLOCK_TYPES

DCCR.JSON_INDENT

DCCR.LOG_STATES

DCCR.n_colors

DCCR.pubu_9

DCCR.pubu_9_raw

DCCR.ylgn_9

DCCR.ylgn_9_raw

Docs

Documentation for all internal names are listed below.

DCCR.BLOCK_TYPES — Constant

The names of the blocks that are encountered during L2 experiments.

DCCR.JSON_INDENT — Constant

Constant for pretty indentation spacing in JSON files.

DCCR.LOG_STATES — Constant

The enumerated states that an L2 logger log can be in.

DCCR.n_colors — Constant

Inferred number of colors used from the color palettes.

DCCR.pubu_9 — Constant

Purple-blue-9 ColorScheme, inferred from the RGB values

DCCR.pubu_9_raw — Constant

Purple-blue-9 raw RGB values, range [0, 1].

DCCR.ylgn_9 — Constant

Yellow-green-9 ColorScheme, inferred from the RGB values.

DCCR.ylgn_9_raw — Constant

Yellow-green-9 raw RGB values, range [0, 1].

DCCR.Agent — Type

L2 agent supertype.

DCCR.DDVFAAgent — Type

DDVFA-based L2 agent.

DCCR.DDVFAAgent — Method

Constructor for a DDVFAAgent using the scenario dictionary and optional DDVFA keyword argument options.

Arguments

opts::AbstractDict: keyword arguments for DDVFA options.

scenario::AbstractDict: l2logger scenario as a dictionary.

DCCR.DDVFAAgent — Method

Creates a DDVFA agent with an empty experience queue.

Arguments

ddvfa_opts::opts_DDVFA: the options struct used to initialize the DDVFA module and set the logging params.

DCCR.Data — Type

Abstract supertype for all Data structs in this library.

DCCR.DataSplit — Type

A basic struct for encapsulating the components of supervised training.

DCCR.DataSplitCombined — Type

A struct for combining training and validation data, containing only train and test splits.

DCCR.DataSplitCombined — Method

Constructs a DataSplitCombined from an existing DataSplit by consolidating the training and validation data.

Arguments

data::DataSplit: the DataSplit struct for consolidating validation Features and Labels into the training data.

DCCR.DataSplitIndexed — Type

A struct for encapsulating the components of supervised training in vectorized form.

DCCR.Experience — Type

Experience block for an agent.

Taken from l2logger_template.

DCCR.Experience — Method

Constructs an Experience, setting the update_model field based upon the block type.

Arguments

task_name::AbstractString: the name of the current task.

seq_nums::SequenceNums: the block and experience number of the Experience.

block_type::AbstractString: the block type ∈ ["train", "test"]. Using "train" sets update_model to true, "test" to false.

DCCR.ExperienceQueue — Type

Alias for a queue of Experiences.

DCCR.ExperienceQueueContainer — Type

Container for the ExperienceQueue and some statistics about it.

DCCR.ExperienceQueueContainer — Method

Creates a queue of Experiences from the scenario dictionary.

Arguments

scenario_dict::AbstractDict: the scenario dictionary.

DCCR.ExperienceQueueContainer — Method

Creates an empty ExperienceQueueContainer with an empty queue and zeroed stats.

DCCR.Features — Type

Definition of features as a matrix of floating-point numbers of dimension (featuredim, nsamples).

DCCR.LabeledDataset — Type

A single dataset of Features, Targets, and human-readable string Labels.

DCCR.LabeledDataset — Method

A constructor for a LabeledDataset that merges two other LabeledDatasets.

Arguments

d1::LabeledDataset: the first LabeledDataset to consolidate.

d2::LabeledDataset: the second LabeledDataset to consolidate.

DCCR.Labels — Type

Definition of labels as a vector of strings.

DCCR.MatrixData — Type

Abstract type for Data structs that represent features as matrices.

DCCR.SequenceNums — Type

Sequence numbers for a block and Experience.

Taken from l2logger_template.

DCCR.StatsDict — Type

Alias for a statistics dictionary being string keys mapping to any object.

DCCR.Targets — Type

Definition of targets as a vector of integers.

DCCR.VectorLabeledDataset — Type

A single dataset of vectored labeled data with Features, Targets, and human-readable string Labels.

DCCR.VectoredData — Type

Abstract type for Data structs that represent features as vectors of matrices.

Base.show — Method

Overload of the show function for DDVFAAgent.

Arguments

io::IO: the current IO stream.

cont::DDVFAAgent: the DDVFAAgent to print/display.

Base.show — Method

Overload of the show function for ExperienceQueueContainer.

Arguments

io::IO: the current IO stream.

cont::ExperienceQueueContainer: the ExperienceQueueContainer to print/display.

Base.show — Method

Overload of the show function for ExperienceQueue.

Arguments

io::IO: the current IO stream.

cont::ExperienceQueueContainer: the ExperienceQueueContainer to print/display.

DCCR.evaluate_agent! — Method

Evaluates a single agent on a single experience, training or testing as needed.

Arguments

agent::Agent: the Agent to evaluate.

exp::Experience: the Experience to use for training/testing.

DCCR.fields_to_dict! — Method

Adds entry to a dictionary from a struct with fields.

Meant to be used with StatsDict.

Arguments

dict::AbstractDict: the StatsDict dictionary to add entries to.

opts::Any: a struct containing fields, presumably of options, to add as key-value entries to the dict.

DCCR.get_index_from_name — Method

Gets an integer index of where a string name appears in a list of strings.

Arguments

labels::Vector{T} where T <: AbstractString: the list of strings to search.

name::AbstractString: the name to search for in the list of labels.

DCCR.initialize_exp_queue! — Method

Initializes an ExperienceQueueContainer from the provided scenario dictionary.

Arguments

eqc::ExperienceQueueContainer: the container with the queue and stats to initialize.

scenario_dict::AbstractDict: the dictionary with the scenario regimes and block types.

DCCR.is_complete — Method

Checks if the Agent is done with its scenario queue.

Arguments

agent::Agent: the agent to test scenario completion on.

DCCR.json_load — Method

Loads the JSON file, interpreted as a dictionary.

Arguments

filepath::AbstractString: the full file name (with path) to load.

DCCR.json_save — Method

Saves the dictionary to a JSON file.

Arguments

filepath::AbstractString: the full file name (with path) to save to.

dict::AbstractDict: the dictionary to save to the file.

DCCR.log_data — Method

Logs data from an L2 Experience.

Arguments

data_logger::PyObject: the l2logger DataLogger.

exp::Experience: the Experience that the Agent just processed.

results::Dict: the results from the Agent's Experience.

status::AbstractString: string expressing if the Experience was processed.

DCCR.run_scenario — Method

Runs an agent's scenario.

Arguments

agent::Agent: a struct that contains an Agent and scenario.

data_logger::PyObject: a l2logger object.

DCCR.sanitize_block_type — Method

Sanitize the selected block type against the BLOCK_TYPES constant.

Arguments

block_type::AbstractString: the selected block type.

DCCR.sanitize_in_list — Method

Sanitizes a selection within a list of acceptable options.

Arguments

selection_type::AbstractString: a string describing the option in case it is misused.

selection::Any: a single selection from a list.

DCCR.sanitize_log_state — Method

Sanitize the selected log state against the LOG_STATES constant.

Arguments

log_state::AbstractString: the selected log state.