Developer Index

This page lists the types and functions that are internal to the DCCR package. Because they are not part of the public API, these names might change relatively frequently between versions and so should not be relied upon.

All internal names are listed in the Index, and each of these entries link to the docstrings in the Docs section.

Index

This section contains a list of internal names that link to their corresponding Documentation.

Methods

DCCR._save_dccr_fig

DCCR._save_dccr_table

DCCR.collect_activations

DCCR.collect_all_activations

DCCR.collect_all_activations_labeled

DCCR.create_accuracy_groupedbar

DCCR.create_boxplot

DCCR.create_comparison_groupedbar

DCCR.create_complex_condensed_plot

DCCR.create_complex_condensed_plot_alt

DCCR.create_condensed_plot

DCCR.create_confusion_heatmap

DCCR.create_custom_confusion_heatmap

DCCR.create_inverted_boxplot

DCCR.df_column_to_matrix

DCCR.dist_exp_parse

DCCR.evaluate_agent!

DCCR.exp_parse

DCCR.feature_preprocess

DCCR.fields_to_dict!

DCCR.get_accuracies

DCCR.get_argparsesettings

DCCR.get_confusion

DCCR.get_deindexed_data

DCCR.get_dist

DCCR.get_index_from_name

DCCR.get_indexed_data

DCCR.get_manual_split

DCCR.get_n_categories

DCCR.get_normalized_confusion

DCCR.get_orbit_names

DCCR.get_tt_accuracies

DCCR.handle_display

DCCR.initialize_exp_queue!

DCCR.is_complete

DCCR.json_load

DCCR.json_save

DCCR.load_default_orbit_data

DCCR.load_opts

DCCR.load_orbits

DCCR.load_sim_opts

DCCR.load_sim_results

DCCR.log_data

DCCR.pack_data

DCCR.packed_dir

DCCR.paper_results_dir

DCCR.permuted

DCCR.results_dir

DCCR.run_scenario

DCCR.safe_unpack

DCCR.sanitize_block_type

DCCR.sanitize_in_list

DCCR.sanitize_log_state

DCCR.save_dccr

DCCR.shuffle_orbits

DCCR.shuffled_mc

DCCR.sigmoid

DCCR.unpack_data

DCCR.unpacked_dir

DCCR.unsupervised_mc

Types

DCCR.Agent

DCCR.DDVFAAgent

DCCR.DDVFAAgent

DCCR.DDVFAAgent

DCCR.Data

DCCR.DataSplit

DCCR.DataSplitCombined

DCCR.DataSplitCombined

DCCR.DataSplitIndexed

DCCR.Experience

DCCR.Experience

DCCR.ExperienceQueue

DCCR.ExperienceQueueContainer

DCCR.ExperienceQueueContainer

DCCR.ExperienceQueueContainer

DCCR.Features

DCCR.LabeledDataset

DCCR.LabeledDataset

DCCR.Labels

DCCR.MatrixData

DCCR.ParsedArgs

DCCR.SequenceNums

DCCR.StatsDict

DCCR.Targets

DCCR.VectorLabeledDataset

DCCR.VectoredData

Constants

DCCR.ARG_CLASS_LABELS

DCCR.ARG_DATA_MATRIX

DCCR.ARG_DATA_SPLIT

DCCR.ARG_Y

DCCR.ARG_Y_HAT

DCCR.ARG_Y_HAT_TRAIN

DCCR.BLOCK_TYPES

DCCR.DOC_ARG_SAVE_DIR

DCCR.JSON_INDENT

DCCR.LOG_STATES

DCCR.SAVE_MAP

DCCR.n_colors

DCCR.pubu_9

DCCR.pubu_9_raw

DCCR.ylgn_9

DCCR.ylgn_9_raw

Docs

Documentation for all internal names are listed below.

DCCR.ARG_CLASS_LABELS — Constant

Common docstring, argument for the class labels as strings used for plot axes.

DCCR.ARG_DATA_MATRIX — Constant

Common docstring, argument for a set of features as a 2-D matrix.

DCCR.ARG_DATA_SPLIT — Constant

Common docstring, argument for the DataSplit used for training, plotting, etc.

DCCR.ARG_Y — Constant

Common docstring, argument for the true target values.

DCCR.ARG_Y_HAT — Constant

Common docstring, argument for the classifier's target outputs.

DCCR.ARG_Y_HAT_TRAIN — Constant

Common docstring, argument for the target estimates on the training data.

DCCR.BLOCK_TYPES — Constant

The names of the blocks that are encountered during L2 experiments.

DCCR.DOC_ARG_SAVE_DIR — Constant

Common doc: directory string to save to.

DCCR.JSON_INDENT — Constant

Constant for pretty indentation spacing in JSON files.

DCCR.LOG_STATES — Constant

The enumerated states that an L2 logger log can be in.

DCCR.SAVE_MAP — Constant

Dictionary mapping the names of result save types to the private wrapper functions that implement them.

DCCR.n_colors — Constant

Inferred number of colors used from the color palettes.

DCCR.pubu_9 — Constant

Purple-blue-9 ColorScheme, inferred from the RGB values

DCCR.pubu_9_raw — Constant

Purple-blue-9 raw RGB values, range [0, 1].

DCCR.ylgn_9 — Constant

Yellow-green-9 ColorScheme, inferred from the RGB values.

DCCR.ylgn_9_raw — Constant

Yellow-green-9 raw RGB values, range [0, 1].

DCCR.Agent — Type

L2 agent supertype.

DCCR.DDVFAAgent — Type

DDVFA-based L2 agent.

DCCR.DDVFAAgent — Method

Constructor for a DDVFAAgent using the scenario dictionary and optional DDVFA keyword argument options.

Arguments

opts::AbstractDict: keyword arguments for DDVFA options.

scenario::AbstractDict: l2logger scenario as a dictionary.

DCCR.DDVFAAgent — Method

Creates a DDVFA agent with an empty experience queue.

Arguments

ddvfa_opts::opts_DDVFA: the options struct used to initialize the DDVFA module and set the logging params.

DCCR.Data — Type

Abstract supertype for all Data structs in this library.

DCCR.DataSplit — Type

A basic struct for encapsulating the components of supervised training.

DCCR.DataSplitCombined — Type

A struct for combining training and validation data, containing only train and test splits.

DCCR.DataSplitCombined — Method

Constructs a DataSplitCombined from an existing DataSplit by consolidating the training and validation data.

Arguments

data::DataSplit: the DataSplit struct for consolidating validation Features and Labels into the training data.

DCCR.DataSplitIndexed — Type

A struct for encapsulating the components of supervised training in vectorized form.

DCCR.Experience — Type

Experience block for an agent.

Taken from l2logger_template.

DCCR.Experience — Method

Constructs an Experience, setting the update_model field based upon the block type.

Arguments

task_name::AbstractString: the name of the current task.

seq_nums::SequenceNums: the block and experience number of the Experience.

block_type::AbstractString: the block type ∈ ["train", "test"]. Using "train" sets update_model to true, "test" to false.

DCCR.ExperienceQueue — Type

Alias for a queue of Experiences.

DCCR.ExperienceQueueContainer — Type

Container for the ExperienceQueue and some statistics about it.

DCCR.ExperienceQueueContainer — Method

Creates a queue of Experiences from the scenario dictionary.

Arguments

scenario_dict::AbstractDict: the scenario dictionary.

DCCR.ExperienceQueueContainer — Method

Creates an empty ExperienceQueueContainer with an empty queue and zeroed stats.

DCCR.Features — Type

Definition of features as a matrix of floating-point numbers of dimension (featuredim, nsamples).

DCCR.LabeledDataset — Type

A single dataset of Features, Targets, and human-readable string Labels.

DCCR.LabeledDataset — Method

A constructor for a LabeledDataset that merges two other LabeledDatasets.

Arguments

d1::LabeledDataset: the first LabeledDataset to consolidate.

d2::LabeledDataset: the second LabeledDataset to consolidate.

DCCR.Labels — Type

Definition of labels as a vector of strings.

DCCR.MatrixData — Type

Abstract type for Data structs that represent features as matrices.

DCCR.ParsedArgs — Type

Type alias for how parsed arguments are treated.

DCCR.SequenceNums — Type

Sequence numbers for a block and Experience.

Taken from l2logger_template.

DCCR.StatsDict — Type

Alias for a statistics dictionary being string keys mapping to any object.

DCCR.Targets — Type

Definition of targets as a vector of integers.

DCCR.VectorLabeledDataset — Type

A single dataset of vectored labeled data with Features, Targets, and human-readable string Labels.

DCCR.VectoredData — Type

Abstract type for Data structs that represent features as vectors of matrices.

Base.show — Method

Overload of the show function for DDVFAAgent.

Arguments

io::IO: the current IO stream.

cont::DDVFAAgent: the DDVFAAgent to print/display.

Base.show — Method

Overload of the show function for ExperienceQueueContainer.

Arguments

io::IO: the current IO stream.

cont::ExperienceQueueContainer: the ExperienceQueueContainer to print/display.

Base.show — Method

Overload of the show function for ExperienceQueue.

Arguments

io::IO: the current IO stream.

cont::ExperienceQueueContainer: the ExperienceQueueContainer to print/display.

DCCR._save_dccr_fig — Method

Wrapper for how figures are saved in the DCCR project.

Arguments

fig: the figure object to save.

dir::AbstractString: the directory to save to.

DCCR._save_dccr_table — Method

Wrapper for how tables are saved in the DCCR project.

Arguments

table: the table object to save.

dir::AbstractString: the directory to save to.

DCCR.collect_activations — Method

Returns the activations from a single directory.

Arguments

data_dir::AbstractString: the single data directory to load the features from.

DCCR.collect_all_activations — Method

Return just the yolo activations from a list of data directories.

Arguments

data_dirs::AbstractArray: the data directories to load the yolo activations from.

cell::Integer: the number of cells corresponding to the windowing procedure.

DCCR.collect_all_activations_labeled — Method

Return the yolo activations, training targets, and condensed labels list from a list of data directories.

Arguments

data_dirs::Vector{String}: the directories to load the data from.

cell::Integer: the number of cells to use in the windowed averaging procedure.

DCCR.create_accuracy_groupedbar — Method

Return a grouped bar chart with class accuracies.

Arguments

data::DataSplit: the original dataset with a train, val, and test split.

y_hat_train::IntegerVector: the classifier estimates from the training data.

y_hat::IntegerVector: the approximated targets generated by the classifier.

class_labels::Vector{String}: the string labels to use for the plot axes.

percentages::Bool=false: optional, flag to use the custom percentage formatter or not.

DCCR.create_boxplot — Method

Return a colored and formatted boxplot of the data.

Arguments

data::RealMatrix: the data as a 2-D matrix of real values.

class_labels::Vector{String}: the string labels to use for the plot axes.

percentages::Bool=false: optional, flag to use the custom percentage formatter or not.

bounds::Tuple{Float, Float}=(0.45, 1.0): optional, the bounds for the y-lim bounds of the plot.

violin_bandwidth::Real=0.01: the bandwidth parameter passed to the violin plot.

DCCR.create_comparison_groupedbar — Method

Return a grouped bar chart with comparison bars.

Arguments

data::DataSplit: the original dataset with a train, val, and test split.

y_hat_val::IntegerVector: the classifier estimates from the validation data.

y_hat::IntegerVector: the approximated targets generated by the classifier.

class_labels::Vector{String}: the string labels to use for the plot axes.

percentages::Bool=false: optional, flag to use the custom percentage formatter or not.

extended::Bool=false: if the plot needs to be extended to another category, compensating for misclassification.

``

DCCR.create_complex_condensed_plot — Function

Create and return a complex condensed scenario plot.

DCCR.create_complex_condensed_plot_alt — Function

Create and return an alternate complex condensed scenario plot.

DCCR.create_condensed_plot — Function

Create and return a simplified condensed scenario plot.

DCCR.create_confusion_heatmap — Method

Returns a handle to a labeled and annotated heatmap plot of the confusion matrix.

Arguments

class_labels::Vector{String}: the string labels to use for the plot axes.

y::IntegerVector: the true targets as integers.

y_hat::IntegerVector: the approximated targets generated by the classifier.

DCCR.create_custom_confusion_heatmap — Method

Returns a handle to a labeled and annotated heatmap plot of the confusion matrix.

Arguments

class_labels::Vector{String}: the string labels to use for the plot axes.

norm_cm::RealMatrix: the normalized confuction matrix to plot as a heatmap.

DCCR.create_inverted_boxplot — Method

Return a colored and formatted boxplot of the data.

DCCR.df_column_to_matrix — Method

Convert a column of lists in a DataFrame into a matrix for analysis.

Arguments

df::DataFrame: the DataFrame containing the column of lists.

row::Symbol: the symbolic name of the row in the DataFrame to convert into a matrix.

DCCR.dist_exp_parse — Function

Parses the command line for common options in distributed experiments.

Arguments

description::AbstractString: optional positional, the script description for the parser

DCCR.evaluate_agent! — Method

Evaluates a single agent on a single experience, training or testing as needed.

Arguments

agent::Agent: the Agent to evaluate.

exp::Experience: the Experience to use for training/testing.

DCCR.exp_parse — Function

Parses the command line for common options in serial (non-distributed) experiments.

Arguments

description::AbstractString: optional positional, the script description for the parser

DCCR.feature_preprocess — Method

Preprocesses one dataset of features, scaling and squashing along the feature axes.

Arguments

dt::ZScoreTransform: the Gaussian statistics of the features.

scaling::Real: the sigmoid scaling parameter.

data::RealMatrix: the 2-D matrix of features to transform.

DCCR.fields_to_dict! — Method

Adds entry to a dictionary from a struct with fields.

Meant to be used with StatsDict.

Arguments

dict::AbstractDict: the StatsDict dictionary to add entries to.

opts::Any: a struct containing fields, presumably of options, to add as key-value entries to the dict.

DCCR.get_accuracies — Method

Get a list of the percentage accuracies.

Arguments

y::IntegerVector: the target values.

y_hat::IntegerVector: the agent's estimates.

n_classes::Integer: the number of total classes in the test set.

DCCR.get_argparsesettings — Function

Common function for how ArgParseSettings are generated in the project.

Arguments

description::AbstractString: optional positional, the script description for the parser

DCCR.get_confusion — Method

Wrapper method for getting the raw confusion matrix.

Arguments

y::IntegerVector: the target values.

y_hat::IntegerVector: the agent's estimates.

n_classes::Integer: the number of total classes in the test set.

DCCR.get_deindexed_data — Method

Turn a DataSplitIndexed into a DataSplit with the given train/test order.

Arguments

data::DataSplitIndexed: the indexed data to consolidate back into a DataSplit.

order::IntegerVector: the order used by the indexed data for correctly deindexing.

DCCR.get_dist — Method

Get the distribution parameters for preprocessing.

Arguments

data::RealMatrix: a 2-D matrix of features for computing the Gaussian statistics of.

DCCR.get_index_from_name — Method

Gets an integer index of where a string name appears in a list of strings.

Arguments

labels::Vector{T} where T <: AbstractString: the list of strings to search.

name::AbstractString: the name to search for in the list of labels.

DCCR.get_indexed_data — Method

Create a DataSplitIndexed object from a DataSplit.

Arguments

data::DataSplit: the DataSplit to separate into vectors of matrices.

DCCR.get_manual_split — Method

Wrapper of stratifiedobs, returns a manual train/test x/y split from a data matrix and labels using MLDataUtils.

Arguments

data::RealMatrix: the feature data to split into training and testing.

targets::IntegerVector: the labels corresponding to the data to split into training and testing.

DCCR.get_n_categories — Method

Returns both the number of F2 categories and total number of weights per class as two lists.

Arguments

ddvfa::DDVFA: the DDVFA module to calculate the statistics for.

n_classes::Int: the number of target classes that the model was trained upon.

DCCR.get_normalized_confusion — Method

Get the normalized confusion matrix.

Arguments

y::IntegerVector: the target values.

y_hat::IntegerVector: the agent's estimates.

n_classes::Integer: the number of total classes in the test set.

DCCR.get_orbit_names — Method

Map the experiment orbit names to their data directories and plotting class labels.

Arguments

selection::Vector{String}: the selection of labels corresponding to both data directories and plotting labels.

DCCR.get_tt_accuracies — Method

Get two lists of the training and testing accuracies.

Arguments

data::MatrixData: the training and testing dataset, containing a vector of training and testing labels data.train.y and data.test.y.

y_hat_train::IntegerVector: the training estimates.

y_hat::IntegerVector: the agent's estimates.

n_classes::Integer: the number of total classes in the test set.

DCCR.handle_display — Method

Handles the display of plots according to arguments parsed by the script.

Arguments

p::Plots.Plot: the plot handle to display if necessary.

pargs::ParsedArgs: the parsed arguments provided by the script.

DCCR.initialize_exp_queue! — Method

Initializes an ExperienceQueueContainer from the provided scenario dictionary.

Arguments

eqc::ExperienceQueueContainer: the container with the queue and stats to initialize.

scenario_dict::AbstractDict: the dictionary with the scenario regimes and block types.

DCCR.is_complete — Method

Checks if the Agent is done with its scenario queue.

Arguments

agent::Agent: the agent to test scenario completion on.

DCCR.json_load — Method

Loads the JSON file, interpreted as a dictionary.

Arguments

filepath::AbstractString: the full file name (with path) to load.

DCCR.json_save — Method

Saves the dictionary to a JSON file.

Arguments

filepath::AbstractString: the full file name (with path) to save to.

dict::AbstractDict: the dictionary to save to the file.

DCCR.load_default_orbit_data — Method

Loads the default orbit data configuration.

Arguments

data_dir::AbstractString: the relative/absolute directory containing the data.

scaling::Float: the sigmoid scaling parameter, default scaling=2.0

DCCR.load_opts — Method

Loads the provided options YAML file.

Arguments

file::AbstractString: the YAML file to load.

DCCR.load_orbits — Method

Load the orbits data and preprocess the features.

Arguments

data_dir::AbstractString: the top-level data directory.

data_dirs::Vector{String}: the subfolders to load.

scaling::Real: the sigmoidal scaling parameter to use.

DCCR.load_sim_opts — Function

Loads and returns the simulation options from the provided YAML file.

Arguments

file::AbstractString="default.yml": options the file to load from the options directory.

DCCR.load_sim_results — Method

Wrapper for loading simulation results with arbitrarily many fields.

Arguments

data_file::AbstractString: the location of the datafile for loading.

args...: the string names of the files to open.

DCCR.log_data — Method

Logs data from an L2 Experience.

Arguments

data_logger::PyObject: the l2logger DataLogger.

exp::Experience: the Experience that the Agent just processed.

results::Dict: the results from the Agent's Experience.

status::AbstractString: string expressing if the Experience was processed.

DCCR.pack_data — Method

Packs the data under the provided experiment name folder into an LFS-tracked tarball.

Arguments

experiment_name::AbstractString: the name of the file destination to pack from the unpacked_dir to the packed_dir.

DCCR.packed_dir — Method

The packed data directory as a DrWatson-style path function.

Arguments

args...: string arguments a subsequent file or folders.

DCCR.paper_results_dir — Method

DrWatson-style paper results directory.

DCCR.permuted — Method

Runs a single Monte Carlo simulation of training/testing on shuffled samples.

Arguments

d::Dict: a logging dictionary storing simulation parameters.

data::DataSplitIndexed: an indexed train/test split of features and labels.

opts::opts_DDVFA: the options for DDVFA construction.

DCCR.results_dir — Method

DrWatson-style local results directory.

DCCR.run_scenario — Method

Runs an agent's scenario.

Arguments

agent::Agent: a struct that contains an Agent and scenario.

data_logger::PyObject: a l2logger object.

DCCR.safe_unpack — Method

If the provided experiment unpacked directory does not exist, this unpacks it.

Arguments

experiment_name::AbstractString: the name of the file to unpack from the packed_dir to the unpacked_dir.

DCCR.sanitize_block_type — Method

Sanitize the selected block type against the BLOCK_TYPES constant.

Arguments

block_type::AbstractString: the selected block type.

DCCR.sanitize_in_list — Method

Sanitizes a selection within a list of acceptable options.

Arguments

selection_type::AbstractString: a string describing the option in case it is misused.

selection::Any: a single selection from a list.

DCCR.sanitize_log_state — Method

Sanitize the selected log state against the LOG_STATES constant.

Arguments

log_state::AbstractString: the selected log state.

DCCR.save_dccr — Method

Saving function for results in the DCCR project.

This function dispatches to the correct private wrapper saving function via the type option, and the to_paper flag determines if the result is also saved to a secondary location, which is mainly used for also saving the result to the cloud location for the journal paper.

Arguments

type::AbstractString: the type of object being saved (see SAVE_MAP).

object: the object to save as type, whether a figure, table, or something else.

exp_name::AbstractString: the name of the experiment, used for the final saving directories.

save_name::AbstractString: the name of the save file itself.

to_paper::Bool=false: optional, flag for saving to the paper results directory (default false).

DCCR.shuffle_orbits — Method

Shuffles the training orbits data.

Arguments

data::DataSplit: the DataSplit orbits data coming from load_orbits.

DCCR.shuffled_mc — Method

Runs a single Monte Carlo simulation of training/testing on shuffled samples.

Arguments

d::Dict: a logging dictionary storing simulation parameters.

data::DataSplit: a train/test split of features and labels.

opts::opts_DDVFA: the options for DDVFA construction.

DCCR.sigmoid — Method

Returns the sigmoid function on x.

Arguments

x::Real: the float or int to compute the sigmoid function upon.

DCCR.unpack_data — Method

Unpacks data at the provided experiment name tarball into a working directory.

Arguments

experiment_name::AbstractString: the name of the file to unpack from the packed_dir to the unpacked_dir.

DCCR.unpacked_dir — Method

The unpacked data directory as a DrWatson-style path function.

Arguments

args...: string arguments a subsequent file or folders.

DCCR.unsupervised_mc — Method

Runs a single Monte Carlo simulation of supervised training and unsupervised training/testing.

Arguments

d::Dict: a logging dictionary storing simulation parameters.

data::DataSplitCombined: a train/test split of features and labels.

opts::opts_DDVFA: the options for DDVFA construction.