Index

This page lists the core methods and types of the Julia component of the DCCR project.

Index

This section enumerates the names exported by the package, each of which links to its corresponding Documentation.

Modules

DCCR.DCCR

Methods

Types

Constants

DCCR.DCCR_VERSION

Docs

This section lists the documentation for every exported name of the DCCR package.

Modules

DCCR.DCCR — Module

A module encapsulating the experiment driver code for the DCCR project.

Imports

The following names are imported by the package as dependencies:

AdaptiveResonance

ArgParse

Base

ColorSchemes

Core

DataFrames

DataStructures

Dates

DelimitedFiles

DocStringExtensions

DrWatson

JLD2

JSON

Latexify

Logging

MLBase

MLDataUtils

NumericalTypeAliases

Pkg

Plots

Printf

ProgressMeter

PyCall

Random

Reexport

StatsBase

StatsPlots

Exports

The following names are exported and available when using the package:

ABGR

ACTIVATION_FUNCTIONS

ADAPTIVERESONANCE_MODULES

ADAPTIVERESONANCE_VERSION

ADIN99

ADIN99d

ADIN99o

AGray

AGray32

AHSI

AHSL

AHSV

ALCHab

ALCHuv

ALMS

ALab

ALuv

ARGB

ARGB32

ART

ARTMAP

ARTMAP_MODULES

ARTModule

ARTOpts

ART_MODULES

AXYZ

AYCbCr

AYIQ

AbstractAGray

AbstractARGB

AbstractBackend

AbstractDataFrame

AbstractDataTransform

AbstractGray

AbstractGrayA

AbstractHistogram

AbstractLayout

AbstractPlot

AbstractRGB

AbstractRGBA

AbstractRNG

AbstractWeights

AdaptiveResonance

All

AlphaColor

AnalyticWeights

Animation

AsTable

AxyY

BGR

BGRA

Between

BezierCurve

ByRow

CIE1931JV_CMF

CIE1931J_CMF

CIE1931_CMF

CIE1964_CMF

CIE2006_10_CMF

CIE2006_2_CMF

CoefTable

Color

Color3

ColorAlpha

ColorGradient

ColorPalette

ColorTypes

Colorant

ColorantNormed

Colors

Cols

ConvergenceException

CovarianceEstimator

CronbachAlpha

DAM

DCCR_VERSION

DDVFA

DDVFA_METHODS

DE_2000

DE_94

DE_AB

DE_BFD

DE_CMC

DE_DIN99

DE_DIN99d

DE_DIN99o

DE_JPC79

DIN99

DIN99A

DIN99d

DIN99dA

DIN99o

DIN99oA

DVFA

DataConfig

DataFrame

DataFrameRow

DataFrames

ECDF

FAM

FancyNumberFormatter

Formatted

Fractional

FrequencyWeights

FuzzyART

GR

GammaNormalizedFuzzyART

Gray

Gray24

GrayA

GroupedDataFrame

HSB

HSI

HSIA

HSL

HSLA

HSV

HSVA

Histogram

InvertedIndex

InvertedIndices

KW

L1dist

L2dist

LCHab

LCHabA

LCHuv

LCHuvA

LMS

LMSA

Lab

LabA

Latexify

Linfdist

Luv

LuvA

MATCH_FUNCTIONS

MSC

MersenneTwister

Missing

MissingException

Missings

Not

OHLC

@P_str

PlotTheme

PlotThemes

PlotUtils

Plots

PrettyTables

Printf

ProbabilityWeights

RGB

RGB1

RGB24

RGB4

RGBA

RGBX

Random

RandomDevice

RecipeData

RecipesBase

RegressionModel

SFAM

Segments

Shape

SimpleCovariance

StatisticalModel

StatsBase

StyledNumberFormatter

SubDataFrame

Surface

Tables

TaskLocalRNG

Transparent3

TransparentColor

TransparentGray

TransparentRGB

UPDATE_FUNCTIONS

UnitRangeTransform

UnitWeights

Weights

XRGB

XYZ

XYZA

Xoshiro

YCbCr

YCbCrA

YIQ

YIQA

ZScoreTransform

adapted_grid

add_theme

addcounts!

adjr2

adjr²

aggregate

aic

aicc

aliases

allcombinations

allowmissing

allowmissing!

alpha

alphacolor

@animate

animate

annotate!

antijoin

areaplot

areaplot!

arrow

artscene_filter

attr!

auto_display

autocor

autocor!

autocov

autocov!

aweights

backend

backend_name

backend_object

backends

bar

bar!

barh

barh!

barhist

barhist!

base_color_type

base_colorant_type

bbox

bic

bitrand

blue

boxplot

boxplot!

brush

by

ccolor

center

cgrad

chroma

cie_color_match

classify

closeall

coalesce

coef

coefnames

coeftable

colmetadata

colmetadata!

colmetadatakeys

color

color_list

color_type

coloralpha

@colorant_str

colordiff

colormap

colormatch

columnindex

combine

comp1

comp2

comp3

comp4

comp5

competerank

complement_code

completecases

confint

contour

contour!

contour3d

contour3d!

contourf

contourf!

cooksdistance

coords

copy_to_clipboard

cor

cor2cov

corkendall

corspearman

counteq

countmap

countne

counts

cov

cov2cor

cronbachalpha

crosscor

crosscor!

crosscov

crosscov!

crossentropy

crossjoin

crossmodelmatrix

cumulant

current

curve_points

curves

curves!

cvec

data_setup!

default

default_cgrad

delete!

deletecolmetadata!

deletemetadata!

denserank

density

density!

describe

deuteranopic

deviance

disallowmissing

disallowmissing!

distinguishable_colors

diverging_palette

dof

dof_residual

dropmissing

dropmissing!

ecdf

emptycolmetadata!

emptymetadata!

emptymissing

entropy

eweights

fillcombinations

fit

fit!

fitted

flatten

font

frame

fweights

gamutmax

gamutmin

gaston

genmean

genvar

geomean

get_W

get_color_palette

get_data_characteristics

get_default

@gif

gif

gkldiv

gr

gray

green

grid

groupby

groupcols

groupindices

gui

harmmean

hdf5

heatmap

heatmap!

hex

hexbin

hexbin!

histogram

histogram!

histogram2d

histogram2d!

hline

hline!

hspan

hspan!

hue

indexmap

indicatormat

informationmatrix

inline

innerjoin

insertcols

insertcols!

inspectdr

inverse_rle

invisible

iqr

isdark

isfitted

islinear

ismissing

iter_segments

kldivergence

kurtosis

latexalign

latexarray

@latexdefine

latexequation

@latexify

latexify

latexinline

latexraw

@latexrecipe

@latexrun

latextabular

@layout

leftjoin

leftjoin!

lens!

levels

levelsmap

leverage

linear_normalization

loglikelihood

mad

mad!

mapc

mapcols

mapcols!

mapreducec

maxad

md

mdtable

mean

mean!

mean_and_cov

mean_and_std

mean_and_var

meanad

meanresponse

median

median!

mesh3d

mesh3d!

metadata

metadata!

metadatakeys

middle

midpoints

missing

missings

mode

model_response

modelmatrix

modes

moment

mov

mp4

msd

mss

ncol

nobs

nonmissingtype

nonunique

norepeats

normalize_hue

nquantile

nrow

nulldeviance

nullloglikelihood

ohlc

ohlc!

optimize_datetime_ticks

optimize_ticks

opts_DAM

opts_DDVFA

opts_DVFA

opts_FAM

opts_FuzzyART

opts_GammaNormalizedFuzzyART

opts_SFAM

order

ordinalrank

outerjoin

pacf

pacf!

pairwise

pairwise!

palette

parametric_colorant

partialcor

passmissing

path3d

path3d!

percentile

percentilerank

performance

pgfplots

pgfplotsx

pie

pie!

plot

plot!

plot3d

plot3d!

plot_color

plotarea

plotattr

plotly

plotlyjs

plots_heatmap

plots_heatmap!

png

portfoliocomposition

portfoliocomposition!

predict

predict!

@printf

proportionmap

proportions

proprow

protanopic

psnr

pweights

pyplot

pythonplot

quantile

quantile!

quantilerank

quiver

quiver!

r2

rand!

randcycle

randcycle!

randexp

randexp!

randn!

randperm

randperm!

randstring

randsubseq

randsubseq!

@recipe

red

reducec

rename

rename!

render

renyientropy

repeat!

reset_default

resetfontsizes

residuals

response

responsename

rgb_string

rgba_string

rightjoin

rle

rmsd

rotate

rotate!

rownumber

rss

r²

sample

sample!

samplepair

savefig

scalefontsize

scalefontsizes

scatter

scatter!

scatter3d

scatter3d!

scatterhist

scatterhist!

scattermat

scattermat_zm

scattermatm

score

select

select!

sem

semijoin

sequential_palette

@series

set_default

set_theme

shape_coords

@shorthands

showtheme

showtheme!

shuffle

shuffle!

skewness

skipmissings

span

@sprintf

spy

spy!

sqL2dist

standardize

std

stderror

stephist

stephist!

sticks

sticks!

stroke

subset

subset!

sum

summarystats

surface

surface!

test_examples

text

theme

theme_palette

tiedrank

title!

totalvar

train!

transform

transform!

translate

translate!

trim

trim!

trimvar

tritanopic

twinx

twiny

unicodeplots

unique!

unstack

@userplot

uweights

valuecols

values

var

variation

vcov

violin

violin!

vline

vline!

vspan

vspan!

webm

weighted_color_mean

weights

whitebalance

winsor

winsor!

wireframe

wireframe!

with

wmedian

wquantile

wrap

wsample

wsample!

wsum

wsum!

xaxis!

xerror

xerror!

xflip!

xgrid!

xlabel!

xlims

xlims!

xticks

xticks!

xyY

xyYA

yaxis!

yerror

yerror!

yflip!

ygrid!

ylabel!

ylims

ylims!

yticks

yticks!

zaxis!

zerror

zerror!

zflip!

zgrid!

zlabel!

zlims

zlims!

zscale

zscore

zscore!

zticks

zticks!

Functions

Types

Constants

DCCR.DCCR_VERSION — Constant

A constant that contains the version of the installed DCCR package.

This value is computed at compile time, so it may be used to programmatically verify the version of OAR that is installed in case a compat entry in your Project.toml is missing or otherwise incorrect.